Add connection fingerprint
This a kind of a weak way to verify that the are connected to the correct peer. Before removing bonjour once connection was established we were sure that aut and runner app were on the same device. The new implementation has a potential race condition on the tcp port assignment: if multiple simulators (which share the same ip) start at the same time and the socket we use to find open ports picks the same ports we could potentially end with the same port being used on different simulators. To be absolutely sure that the peer is the intended one we double check after connection is up by comparing a uuid that is generated client side